add support for HTTPS with Cert Validation through OpenSSL - add TR-181 control
RDKB-4778 Implement TR-181 Parameter to enable/disable SSL certificate check to ACS server

DESCRIPTION:
- Adding the common library support here for the changes in ccsp-tr069-pa for ssl certificate check to ACS.
- Implement a new TR181 parameter "Device.ManagementServer.X_LGI_COM_ValidateManagementServerCertificate".
- Parameter decides whether SSL certificate check for ACS server will be enabled or disabled.

TESTING PERFORMED:
1) Configure TR069 parameters for connecting to ACS server.
2) Use "HTTPS" in Device.ManagementServer.URL in order to establish SSL connection to ACS server.
3) If the parameter X_LGI_COM_ValidateManagementServerCertificateID is set to True
the modem is failing to establish TR69 connection to ACS server because it lacks related/proper certificates to verify ACS.
4) If the parameter X_LGI_COM_ValidateManagementServerCertificateID is set to False.
the modem is establishing TR69 connection to ACS server which it will establish anyway
when no check is happening as the parameter is set to False.
5) If the modem is having the necessary/proper certificate, then it is establishing
TR69 connection to ACS server irrespective of whether the parameter is set True or False.

Signed-off-by: Sandipan Sinha <ssinha.contractor@libertyglobal.com>