also wait for NewSocketEvent in AnscDetoStop()
Fix for crash seen on SPV of:

  Device.ManagementServer.X_LGI-COM_ConnectionRequestPort

which calls "Cancel" and "Engage" on TCP server.

Fix by preventing AnscDetoRecvTask() from acquiring the
NewSocketEvent lock and therefore using the object which is about to
be destroyed in the cleanup procedure (the crash was seen in the call
to AnscResetEvent(&pMyObject->NewSocketEvent)).

AnscDetoStop() already used AnscWaitEvent() to acquire and hold
RecvEvent and SendEvent, but NewSocketEvent seems to be needed too.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>