limit wait time in case of registerCapabilities failure ( OFW-4866 )
[Issue]
On CCSP component start-up, CcspBaseIf_registerCapabilities is called to register the component with CR.
An issue was seen with LMLite where this call kept failing continuously and after each failure the waiting time was doubled.
After a few failures, new attempts will happen after more than a day of waiting.

[Solution]
Limit the waiting time for retries to 80 seconds to make debugging easier when the issue is reproduced.