fix CcspBaseIf_registerCapabilities_rbus() timeout error handling ( OFW-5342 CH7465OFW-1523 )
Fix TR069 and Plume agents not working after reboot.

[Issue]
A random issue occurs during bootup that causes TR69 to stay down
because certain components aren't able to register with the CR.

The components aren't able to register when
CcspBaseIf_registerCapabilities_rbus fails. The failure occurs when
rbus_addElement, which is called for every datamodel parameter a
component has, returns an error because of a timeout after 6 seconds.
When this timeout occurs, rbus_removeElement is called to remove all
added elements upto but not including the element that triggered the
timeout.

In the next iteration when CcspBaseIf_registerCapabilities_rbus is
called, the same element that triggered a timeout will now trigger a "duplicate entry" error
because it wasn't removed earlier.

All future calls to CcspBaseIf_registerCapabilities_rbus will return an
error because of the "duplicate entry" error.

[Solution]
When an error occurs in rbus_addElement, remove all elements up to and
including the element with the error instead of excluding that element.

[Testing performed]
Since the timeout error is difficult to reproduce manually, changes were
made to the code to simulate the error.
When a timeout error was added, the issue was seen and TR69 was unable
to come up.
After introducing the fix and simulating the timeout error again, the
issue was not seen.

[Upstream Candidate]
Yes