RDKBDEV-2305: fix CcspBaseIf_registerCapabilities_rbus() timeout
Reason for change:

fix CcspBaseIf_registerCapabilities_rbus() timeout error handling

Fix TR069 and Plume agents not working after reboot.

[Issue]
A random issue occurs during bootup that causes TR69 to stay down
because certain components aren't able to register with the CR.

The components aren't able to register when
CcspBaseIf_registerCapabilities_rbus fails. The failure occurs when
rbus_addElement, which is called for every datamodel parameter a
component has, returns an error because of a timeout after 6 seconds.
When this timeout occurs, rbus_removeElement is called to remove all
added elements upto but not including the element that triggered the
timeout.

In the next iteration when CcspBaseIf_registerCapabilities_rbus is
called, the same element that triggered a timeout will now trigger a "duplicate entry" error
because it wasn't removed earlier.

All future calls to CcspBaseIf_registerCapabilities_rbus will return an
error because of the "duplicate entry" error.

[Solution]
When an error occurs in rbus_addElement, remove all elements up to and
including the element with the error instead of excluding that element.

Test Procedure: Sanity.
Since the timeout error is difficult to reproduce manually, changes were
made to the code to simulate the error.
When a timeout error was added, the issue was seen and TR69 was unable
to come up.
After introducing the fix and simulating the timeout error again, the
issue was not seen.

Risks: None.
Change-Id: I967bb11d62f2f0f88361ab0b15cc1e9df44f89e1
Signed-off-by: Hamza Ben Zekri <hbenzekri@libertyglobal.com>