fix ACS Connection dies by log "The event 2 PERIODIC is already there" ( OFW-5881 )
[Issue]
In some corner case if we lose the Ip address after binding and before
starting ssl_connect(), the handshake will keep on pending forever and
the ACS connectivity will be hanged waiting for the response that never
comes.

[Solution]
Use non blocking socket instead of a blocking one and implement a
timeout using select to see if we are able to read anything. If not call
error and goes for a retry.

[Testing Performed]
Use MV1, MV2+ and MV3 to check the ACS connectivity using HTTPS URL.

Upstream Candidate: Maybe.

Signed-off-by: Oussema ben abdallah(obenabdallah.ext@libertyglobal.com)