fix modem not reporting telemetry data due to failing of lastrebootreason dml ( OFW-6102 CH7465SI-1167 TG2492SI-626 )
[Issue Description]
As part of ofw dm version 15.0.0, the DM "DeviceInfo.X_RDKCENTRAL-COM_LastRebootReason" has been aliased and made it invisible for ACS.
So when GPV does for the same DM from ACS , it fails and it also impact the GPV for the other set of DMs ( behaviour as per ACS design)

[Solution]
Reverted the aliasing of DeviceInfo.X_RDKCENTRAL-COM_LastRebootReason.

[Testing Performed]
Ensured that GPV of DeviceInfo.X_RDKCENTRAL-COM_LastRebootReason from ACS is working fine.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)